For pending kernel version changes, because security kernel breaks logic for collecting CVEs in unrelease kernels up to the release kernel.
Specifically the security kernel bumped the abi.release numbers and
reverted the previous pending kernels with CVEs in them. However the
remerge of these kernels left the commits and commit messages under
the previousl abi.release number.

  abi-1.release-1 - commits for CVE
  abi.release       emergency kernel
  abi+1.release+1   remerge commits for abi-1.release-1

the scripts do not see the CVEs in abi-1.release-1 as belonging to
the current abi+1.release+1 kernel, which is the kernel that the
fixes will be released in. So manually override

the release mapping is as follows
raring
  3.8.0-19.29  ->  3.8.0-22.33
  3.8.9-20.31  ->  3.8.0-22.33

quantal
  3.5.0-29.49  ->  3.5.0-31.52

precise
  3.2.0-42.67  ->  3.2.0-44.69