Make scripts/check-syntax aware of the Ubuntu CNA pool of CVE numbers
Now scripts/check-syntax will check the Ubuntu CNA pool of CVE numbers,
if the symbolic link for embargoed is set up correctly, and report the CVE
numbers that we have assigned that do not yet have corresponding files
in UCT.

Failing this does not cause check-syntax to fail so it should not prevent
a checkin of other data when used as a bzr plugin. But it will be annoying.

Tested with and without the embargoed symbolic link, but not yet tested
with no entries missing.