CVE-2012-3412: dropped one of the SHA1s from the break-fix
commit 1485348d2424e1131ea42efc033cbd9366462b01 ("tcp: Apply device TSO segment
limit earlier") was revert in 3.2 stable as it was causing connections to stall
when PMTU event occurs.  As per the revert message, this commit is not actually
required to fix the CVE, it just fixes a performance regression on some extreme
cases.  Thus, I'm dropping this SHA1 from the tracker.