sync-bugs-kernel.py: fix breakage due to xenial opening
sync-bugs-kernel.py was breaking on some android only cves because the
first source package queried for did not exist in the cve db entry file.
Once that was fixed, xenial nominations weren't being added due to
querying if 'xenial' existed in the cve file, but of course, xenial is
the devel release and thus the relevant entries are under 'devel'.

Other misc cleanups:
- fixed some pyflakes whinging
- added additional debugging statements
- moved the majority of the main level of the script to be protected
  against a '__MAIN__' check so that the functios of the script could
  be imported into a debugging python interpreter. This led to some
  gratuitous whitespace changes.