CVE-2015-8660: added break commit
Note that this 'break' commit is not really accurate in this case.  This commit
is the commit that introduced ovl in mainline, even if we had it already in our
old kernels.  But these were not really affected by this issue, so this 'break'
commit should make the tracker happy.