sis-generate-usn: look for lts kernel in any of the source packages
With the addition of the kernel meta packages, the test to detect
whether the script was generating a USN for an lts-backport kernel (and
thus should generate the preamble text for it) was no longer working
correctly due to the test looking for a subset. Instead, check if an lts
kernel is any of the source packages.