check-syntax: fix --newer option, then neuter check to perform during --strict
The --newer command line option did not set a default value, and used
the store_true action when --newer was passed. When used to determine
whether or not to check whether versions marked as 'released' are newer
than what the archive shows as the latest, the check was whether the
'opt.newer' field equaled False. Since no default was set, that
condition was never true, and the check was not being done.

I fixed the option, then added a protection to not backtrace and
instead warn about precise/esm released versions not being able to
be found in source_map.py's generated table. This led to a bunch of
flashplugin cves showing up, which I corrected in an earlier commit,
and some nvidia module cves, which are more difficult to address,
due to the fixes being put in place by superceding packages.

Also, with that in place, issuing an update and then committing it to
the UCT tree would automatically generate an error, because the local
pacages_mirror database would not have the information about the just
released update.

So with all that, I've added the 'warn on newer version' check to the
--strict option (which most people don't enable by default), and it can
be disabled from that with the --newer option. This of course could be
revisited, if a way to handle the retired nvidia issues can be put in
place.