check-cves: Support printing CVE descriptions to a command file
This commit allows for the experimental processing mode to prepend "# "
to each line of a CVE description and to print the result to a temporary
command file that's opened in an editor.

There's a known issue with this change in that additional newlines are
being inserted at times, making the CVE descriptions take up more lines
than necessary.