check-cves: Fully process a command file
Fully process add, ignore, and skip commands found in a command file.
This results in CVE files being created, lines being added to
ignored/not-for-us.txt, or a no-op, respectively.

There's a known issue with this commit when a command file contains two
or more add commands for multiple packages corresponding to a single
CVE. Rather than adding all packages to the CVE file, only the last
package is added.