CVE-2017-1000405: Update introducing commit
We found that Precise, which has not been receiving a backport of the commit
which RH claims is introducing the issue, does not reproduce with the POC
given. The commit had been backported to Trusty onwards which have been
fixed now. So add the SHA1 as new breaks.

In Trusty the commit is a SAUCE patch without reference to upstream SHA1.
For that reason also updating the autotriager hint file.