prepare-kernel-usn.py: use glitch fallback for prior usn version
The reason for using this is that for kernels where we don't publish
USNs (e.g. linux-hwe-edge), prepare-kernel-usn.py would report ALL
the CVEs that had ever been addressed since the initial publication,
making it difficult to see what the actual CVEs addressed in this
specific update, if any. We use prepare-kernel-usn.py as part of the
kernel bug signoff process to decide whether a package should go into
-security or not, even if we don't end up publishing a USN for it.

By using the glitchdb fallback, we only search for CVEs back to the last
publication.