sync-from-usns.py: add option to stage for git commit
Add a '-g' option that, when passed with -u, not only updates the CVE
status, but adds the modified file to the git index in preparation for
committing.

(I think the next step would be to just go ahead and perform the git
commit as well, with a formatted commit message giving information
about the USN issued. But more sanity checking of git state would
need to happen first.)

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>