CVE-2015-8539: Unretire kernel keyring issue
Previous attempts at correcting the break info and unretiring this
commit have been unsuccessful due to me not adjusting the individual
status lines to needs-triage. The autotriage bot didn't retriage the
status because everything was marked as needs-triage or ignored and then
the security team eventually retired the CVE since there were no
affected packages.

There's no useful historical data in the individual package status lines
so blow the status away and merge in a clean set of status lines from
active/00boilerplate.linux so that the autotriage bot will do a full run
for all packages and releases with the correct break-fix info.

Fixes: 629dadef0809 ("CVE-2015-8539: Move break commit to start of git history")
Fixes: 2381eb798787 ("CVE-2015-8539: Unretire kernel keyring issue and improve break info")
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>