scripts/html_export.py: Sort packages in output
Sort the packages associate with a CVE by alphabetical order. My primary
interest for this change is to make it so that the linux source package
is listed above any other kernel related source packages. However, it
will also introduce some sanity to non-kernel CVEs since the ordering is
currently "random" due to the existing code cycling through the keys of
a dictionary.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>