scripts/git-hooks: add prepare-commit-msg hook to check on git merge
Sadly, the pre-commit hook does not get run when git merge is invoked,
so adding a prepare-commit-msg hook that should only perform checks when
merges are performed as an alternative approach. Unfortunately, it
leaves the tree in an incomplete merged state if the hook aborts, so
it's not as nice as the pre-commit hook.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>