mariadb-10.1 replaced by mariadb-10.3 in disco.
MaraiaDB 10.1 was replaced by Mariadb 10.3 in Disco. We don't have the
resources to perform a more thorough analysis of which CVEs from 10.1
ought to have been moved to 10.3, but it's a bit acedemic: Either the
MariaDB team will address them in an update or they won't. Our database
won't influence their work. So, remove 10.1 from disco, and we'll add
the mariadb-10.3 stanzas to the relevant CVEs when the MariaDB team
provide us with fixes.