Fixing bug when parse_CVEs consider debian lines
parse_CVEs consider all lines in .changes file to grab the CVE numbers.
The issue is that in cases such as: debian/patches/CVE-2018-190x-*.patch,
it will grab this line as CVE-2018-190, what is not true. Only the
CVE lines should be considered, but kernel team uses a different way
to document changelog, so removing the debian line and considering
the whole text is fine. Also removing obsolte CAN lines.