uct.el: Major updates to allow to filter on various properties
Allow to filter on source (aka directory - active, embargoed, ignored,
retired) or on status (released, needs-triage, needed etc) or to filter
based on regular expression to match against the description - this is
all in addition to the existing filtering based on package name (which
can now be a list of packages).

So can relatively easily find all CVEs against say wordpress with XSS in
description (this then allows to easily find CVEs of interest during CVE
triage etc).