Retriage openjdk-8 CVEs for disco and devel
openjdk-8 now exists in disco and devel (eoan) so fix current check-syntax
warnings about this - for all which where DNE for disco I copied the status
from cosmic -> disco and then did the same for eoan. Finally I reviewed the
changelog for the current openjdk-8 packages in disco and eoan and updated
the few CVEs to show where we had already fixed these with newer
uploads (compared to the version in cosmic).