kernel: Move a number of pending lines back to released
The autotriage bot got confused by the security respin, which was done
in private, for MDS. All of the existing "pending (VERSION)" statuses
which were changed to "released (VERSION)" during USN publication were
changed back to pending by the bot because the VERSION string was
incorrect. The VERSION was incorrect because it had been bumped in the
security PPA, out of view from the bot.

Adjust all of the statuses to released now that the VERSION is correct.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>