uct.el: Use P to filter by package
This frees up p to be used to go to the previous line (to match n for next
line). Since this now stops using P to filter to the current package,
instead use C-u P (universal argument before using P) to filter by the
package of the current CVE. This also allows to simplify the code a bit
too.