source_map: fix source map checking when apt sources are used
When packages-mirror is not set up, the apt sources lists are used for
setting up the sources map; however, it included some cruft dating from
when partner was named commercial and didn't handle the devel release
pockets of devel-security and devel-updates not being populated.

Also add checking to ensure each component (main, restricted, universe,
and multiverse) are there for the main archive.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>