cve-mode: Allow to customise options including how to prefix notes
Seems some folks like to do:

name> this is a really lone note which is spread over more than just a
 single line

Whilst others like:

name> this is a really lone note which is spread over more than just a
name> single line

Well let's support both - so you can set `cve-mode-prefix-notes-lines` to
`first` to get the first case, or `all` to get the second. By default we go
with `first`.

Also allow to customise the default notes user name, and the fill
column (although the default of 76 seems to match what cve_lib uses.)