prepare-kernel-usns.py: pass multiple json files to sis-generate-usn
If multiple ubuntu releases are covered, to avoid potentially the
same binary changes files, we pull down changes by releases and then
pass sis-generate-usn over them all. With the addition of the json
files recording binary versions, we'll need to pass multiple json
files generated to sis-generate-usn, so do that.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>