cve.css: Don't bold the last-child in a table
Commit d865e997fa78 ("Update static files with boostrap styles")
introduced unconditional bolding of the last cell in a row of a table.
I believe that the intent was to highlight the package statuses but it
also mistakenly started bolding the "Fixed by <SHA1>" lines in CVE pages
representing kernel CVEs.

This commit stops bolding the "Fixed by <SHA1>" cells but also stops
unconditionally bolding all of the last cells in a row of a table. It is
a little easier on the eyes, in general, and the HTML/CSS already has
provisions to bold certain cells such as when the status is "needed" or
the priority is "high" or "critical".

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>