14.04 ESM kernels: Move status from pending to released
The 14.04 ESM PPA recently moved and previous 14.04 ESM kernel updates
were not included in the move. Therefore, the kernel CVE autotriage bot
floated the version info to the oldest present kernel updates that it
could find in the new PPA. Move the status from pending back to released
and retain the new version strings.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>