scripts/packages-mirror: Just source ~/.ubuntu-cve-tracker.conf like other scripts do
This is a temporary fix since we don't have python3-configobj on keule.
Ideally no scripts would source this file and instead it would be a
proper config file and be parsed by a ConfigParser or similar but until
that time just source it like all the other old-skool scripts do.