CVE-2019-2215: CVE-2020-0030: Replace break-fix
The fix for CVE-2020-0030 describes a race involving wake_up_poll
introduced by CVE-2019-2215 fix.

The CVE-2019-2215 fix did not introduce a regression, but was an
incomplete fix that is completed by CVE-2020-0030.

By that logic, we also update the break commit for CVE-2019-2215 which
was unspecified as the commit that added the binder driver to the tree.

Then, we use the CVE-2019-2215 fix as the breaks commit for
CVE-2020-0030, making sure that if the incomplete fix is brought in,
that we will know to get the fixup too.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>