cve_lib: handle package_info_overrides when not in $UCT
In some situations, scripts are run from a reviewed directory that
is not a full checkout of UCT, with the $UCT environment variable
set to point at a full check out of UCT data and meta_info. Handle
correctly loading the meta_info package override list json file in
this situation.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>